+++ src/utils/ScoreCollector.java
@@ -3,7 +3,7 @@ package utils;
 public class ScoreCollector implements Collector {
 	
 	float[] weights;
-	public double score;
+	public float score;
 	
 	public ScoreCollector(float[] w) {
 		weights = w;
@@ -16,7 +16,7 @@ public class ScoreCollector implements Collector {
 	}
 
 	@Override
-	public void addEntry(int x, double va) {
+	public void addEntry(int x, float va) {
 		score += weights[x]*va;
 	}
